pyproject.toml
[build-system]
requires = ["flit_core >=3.2,<4"]
build-backend = "flit_core.buildapi"

[project]
name = "firecrest_wflow"
dynamic = ["version", "description"]
authors = [{name = "Chris Sewell", email = "chrisj_sewell@hotmail.com"}]
readme = "README.md"
license = {file = "LICENSE"}
classifiers = [
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
]
requires-python = ">=3.8"
dependencies = [
    "pyfirecrest~=1.2.0",
    "aiohttp~=3.8",
    "aiofiles~=22.1",
    "sqlalchemy==2.0.0rc2",
    "jinja2"
]

[project.optional-dependencies]
cli = ["click", "click-config-file", "pyyaml"]
testing = ["pytest"]

[project.scripts]
fc-wflow-mock = "firecrest_wflow._test:main"
fc-wkflow = "firecrest_wflow.cli:main"

[project.urls]
Home = "https://github.com/chrisjsewell/firecrest-workflows"

[tool.isort]
profile = "black"
force_sort_within_sections = true

[tool.ruff]
line-length = 110
extend-select = ["B0", "C4", "ICN", "ISC", "N", "RUF", "SIM", "T20"]

[tool.mypy]
show_error_codes = true
strict = true

[[tool.mypy.overrides]]
module = [
    "firecrest.*",
    "tomllib.*",
    "yaml.*",
    "click_config_file.*",
]
ignore_missing_imports = true
